+++ data-persistence/data.tf
@@ -1,0 +1,22 @@
+data "aws_caller_identity" "current" {}
+data "aws_region" "current" {}
+
+data "aws_vpc" "application_vpcs" {
+  tags = {
+    Name = "Application VPC"
+  }
+}
+
+data "aws_subnets" "subnet_ids" {
+  filter {
+    name = "tag:Name"
+    values = ["Private application ${data.aws_region.current.name}a subnet",
+    "Private application ${data.aws_region.current.name}b subnet"]
+  }
+}
+
+data "terraform_remote_state" "rds" {
+  backend   = "s3"
+  workspace = var.DEPLOY_NAME
+  config    = local.rds_remote_state_config
+}